requirements.txt
Django==4.1.7
Jinja2==3.1.2
MarkupSafe==2.1.2
Pillow==9.4.0
anyio==3.6.2
asgiref==3.6.0
cachetools==5.3.0
certifi==2022.12.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.2
google-api-core==2.11.0
google-api-python-client==2.81.0
google-auth==2.16.2
google-auth-httplib2==0.1.0
google-auth-oauthlib==1.0.0
googleapis-common-protos==1.58.0
gspread==5.7.2
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httplib2==0.21.0
httpx==0.23.3
hyperframe==6.0.1
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
mpld3==0.5.9
numpy==1.24.2
oauthlib==3.2.2
packaging==23.0
pip==23.0.1
protobuf==4.22.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
python-dateutil==2.8.2
python-telegram-bot==20.1
pytz==2022.7.1
requests==2.28.2
requests-oauthlib==1.3.1
setuptools==67.6.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.3
tzdata==2022.7
uritemplate==4.1.1
urllib3==1.26.15
wheel==0.40.0
